Fix addSex and plotSex
plotSex() returned an error message "Error: all(c("predictedSex", "xMed", "yMed") %in% names(object)) is not TRUE" when running with the examples, where addSex() returned GMsetEx using minfiData example data. It is because the addSex() only adds predictedSex column but not xMed or yMed. I fix addSex(). 

Another bug is the check point of "names(object)". As the object is [Genomic]MethylSet, we should use "colnames(pData(object))".