Add DelayedArray support for compartments() and createCorMatrix()
Support is not-fully optimised but should be reasonable since only a small-ish subset of loci are used.

In `createCorMatrix()`, the subsetted `M` matrix, with dimensions = nloci * nsamples, is realized in-memory. This is relatively small since it is restricted to a single chromosome, typically only includes Open Sea probes, and has SNPs removed.

A complete implementation will require `cor()`, `.returnBinnedMatrix()`, `.getFirstPC()`,  `.fsvd()`, `.meanSmoother()`, and `.unitarize()` to support DelayedArray instances.