Fix IlluminaMethylationManifest objects on-the-fly
The getManifest() methods now replace all the DataFrame instances
contained in the IlluminaMethylationManifest object to return with
DFrame instances. The reason for doing this is that, starting with
S4Vectors 0.24.0, DataFrame instances should no longer be created
or used, and DFrame instances should always be used instead.

Note that this should be a temprary fix only. A better/cleaner fix
would be to update the data in the IlluminaHumanMethylation*manifest
packages.